Consertei alguns erros na tela de Nome
